BB84 protocol works.
The key distribution works. Random key is created but not checked if someone
tried to eavesdrop. The code got its errors though. For example the classical
communication is not fully understood. That causes the code to only work once.
Secondly the number of qubits transfered cant be increased because "there are
no more qubits available"-error is thrown.