The algorithms to generate key are better
Previously alice send all the qubits to Bob and afterwards compairing
the basis and therefore filtering the wrong qubits on both sides.
Now this is process is done one by one to increase the stability of the
script. The previous way of doing it produced a weird problem where alice
would send all of her qubits and then after all have been sent bob would
start. That caused the script to crash since to many qubits were used.